fix: make the tree build on Linux, where it had never been compiled
First GCC build of this repository. Grouped by cause rather than by file:

Two type specifiers only MSVC accepts, both the same grammar rule -- a
functional cast takes one simple-type-specifier, so `struct stat{}` and
`unsigned int{}` are ill-formed. The POSIX half of confined-file.cpp had never
been compiled by anything, so this is the first compiler to see it.

The vendored Luau C++ headers tripped 171 warnings under the project profile:
GCC has no -Weverything, so the per-compiler pragma block has to name each one,
and it named only two. -Wshadow also flags lambda parameters that reuse an
enclosing local's name, which MSVC does not.

lua_objlen returns int and was compared and passed against std::size_t at three
call sites. A redundant std::move under C++23 implicit move, a [[nodiscard]] on
a void function, and one uncalled helper were all fatal under -Werror and
invisible on MSVC.

Two portability bugs the Windows-only gate could not have found. The module
autoloader globs every .cpp under source/, so the ONNX Runtime adapter was
compiled on a platform whose include path never carries its header, against
onnx-engine.hpp's own documented contract; it is guarded now, and the
structural fix -- a manifest key for platform-specific sources -- is recorded
rather than invented during a verification pass.

And a safety check that could not work on POSIX: remove_plain_tree captured a
directory's identity, deleted its children, then compared, but POSIX counts one
link per subdirectory, so removing a child drops the parent's st_nlink and the
function refused its own work. Windows reports 1 for every directory, so it
never fired there. The comparison across a delete now uses an identity without
st_nlink; every other use keeps it, including the hard-link check on files.
Nothing is lost, because rmdir still fails with ENOTEMPTY rather than deleting
anything the scan did not verify.

Verified on both platforms: GATE: PASS on Linux (60 tests) and on Windows
(61 tests; the annotation suite needs jsonschema, which the Linux run predates).

+++ modules/task/source/task/ffi/uf-tables.cpp
@@ -726,7 +726,10 @@ namespace uf::task
                 );
             }
             auto const absolute = lua_absindex(state, index);
-            auto const size     = lua_objlen(state, absolute);
+            // lua_objlen answers a length as int and never a negative one; the
+            // cast is what lets it meet the std::size_t counts below without a
+            // signed/unsigned comparison.
+            auto const size     = static_cast<std::size_t>(lua_objlen(state, absolute));
             auto result         = std::vector<std::string>{};
             result.reserve(size);
             for (auto child = std::size_t{1}; child <= size; ++child)
